requirements.txt
streamlit
torch
torchvision
numpy
opencv-python
matplotlib
pillow
